cleaned up script and pipeline
