+++ eng/pipelines/templates/jobs/archetype-sdk-client.yml
@@ -64,7 +64,7 @@ jobs:
 
       - script: |
           npm install ./eng/tools/versioning
-          node eng/tools/versioning/set-dev-dependencies.js --build-id "$(Build.BuildNumber)" --repo-root "$(Build.SourcesDirectory)" --service "$(folder)"
+          node eng/tools/versioning/set-dev.js --build-id "$(Build.BuildNumber)" --repo-root "$(Build.SourcesDirectory)" --service "$(folder)"
           node common/scripts/install-run-rush.js update
         condition: eq(variables['SetDevVersion'],'true')
         displayName: "Update package versions for dev build"
